requirements.txt
streamlit
pandas
pdfplumber
camelot-py[cv]
openpyxl
python-docx
docxtpl
matplotlib
streamlit-authenticator
pyyaml